remove decimal form race table

+++ db/migrate/20170713135143_create_races.rb
@@ -7,9 +7,9 @@ class CreateRaces < ActiveRecord::Migration[5.1]
       t.belongs_to :category
       t.string :recipients
       t.integer :race_value
-      t.decimal :commission
+      t.integer :commission
       t.integer :pieces_amount
-      t.decimal :compensation_start_amount, :precision => 8, :scale => 2
+      t.integer :compensation_start_amount, :precision => 8, :scale => 2
       t.integer :max_attendees
       t.integer :kind, :default => 1
       t.integer :status
